chore: update DOCKERFILE param for all pipelines to use local file (#60)
When these components were onboarded, the Dockerfile specified in the
HTML form was a full URL pointing to the github repository. Since the
dockerfile will always be local and is part of the source, it should
point to the local file instead. This will also fix a failing check in
the RH Enterprise Contract.

Signed-off-by: Lance Ball <lball@redhat.com>